+++ src/Common/Category/Constant.js
@@ -1,0 +1,43 @@
+import item01 from "./img/item01.svg";
+import item02 from "./img/item02.svg";
+import item03 from "./img/item03.svg";
+import item04 from "./img/item04.svg";
+import item05 from "./img/item05.svg";
+import item06 from "./img/item06.svg";
+import item07 from "./img/item07.svg";
+import item08 from "./img/item08.svg";
+
+export const CATEGORY_LIST = [
+  {
+    img : item01,
+    name: "생활용품"
+  },
+  {
+    img : item02,
+    name: "아이스크림"
+  },
+  {
+    img : item03,
+    name: "탄산음료"
+  },
+  {
+    img : item04,
+    name: "유제품/커피 음료"
+  },
+  {
+    img : item05,
+    name: "생수/과채/기타음료"
+  },
+  {
+    img : item06,
+    name: "간편식사"
+  },
+  {
+    img : item07,
+    name: "빵/과자류"
+  },
+  {
+    img : item08,
+    name: "기타"
+  }
+]
